+++ src/modules/comments.test.js
@@ -3,7 +3,7 @@
  * @jest-environment jsdom
  */
 
-import { countComment } from './operation.js';
+import { countComments } from './comments.js';
 
 test('Count and Display Number of Comments', () => {
   document.body.innerHTML = `
@@ -11,7 +11,7 @@ test('Count and Display Number of Comments', () => {
     <div class="comments"></div>
     <div class="comments"></div>
     <div class="comments"></div>`;
-  countComment();
+  countComments();
   const amount = document.querySelector('.comments-amount').textContent;
   expect(amount).toBe('(3)');
 });
